\name{compute_newStrainLevel}
\alias{compute_newStrainLevel}
\title{Compute the newStrainLevel}
\usage{
  compute_newStrainLevel(ss, scenario_spec)
}
\arguments{
  \item{ss}{A solved system as produced by run_system}

  \item{scenario_spec}{The specification of the scenario}
}
\description{
  Mostly for testing
}

