\name{compute_offThreshold}
\alias{compute_offThreshold}
\title{Compute the approximate offThreshold - the value below which strains are set to extinct}
\usage{
  compute_offThreshold(ss, scenario_spec)
}
\arguments{
  \item{ss}{A solved system as produced by run_system}

  \item{scenario_spec}{The specification of the scenario}
}
\description{
  Mostly for testing
}

