\name{compute_parameters}
\alias{compute_parameters}
\title{This function transforms the input parameters into a format that the equations can use}
\usage{
  compute_parameters(constants, scenario_parameters)
}
\arguments{
  \item{constants}{The parameters that are unlikely to
  change from scenario to scenario, as produced by
  get_de_defaults}

  \item{scenario_parameters}{The parameters that specify
  the scenario}
}
\description{
  This function transforms the input parameters into a
  format that the equations can use
}

