\name{eventFunc}
\alias{eventFunc}
\title{The event function. Manipulates the state and parameter functions of thee system}
\usage{
  eventFunc(t, y, parms)
}
\arguments{
  \item{t}{time in the system}

  \item{y}{state variables}

  \item{parms}{The extra system parameters}
}
\description{
  The event function. Manipulates the state and parameter
  functions of thee system
}

