\name{format_data}
\alias{format_data}
\title{Names and reshapes the data so that it is easy to plot}
\usage{
  format_data(solved_system)
}
\arguments{
  \item{solved_system}{The output from ode that needs to be
  formatted}
}
\description{
  Names and reshapes the data so that it is easy to plot
}

