\name{get_all_scenarios}
\alias{get_all_scenarios}
\title{Returns all the pre-specified scenarios}
\usage{
  get_all_scenarios()
}
\description{
  This function is a poor way of storing pre-specified
  scenarios. This function is mainly a placeholder for a
  mechanism that will store specified scenarios in a DB
}

