\name{get_scenario}
\alias{get_scenario}
\title{Retrieves one of the pre-specified scenario specifications}
\usage{
  get_scenario(scenario_name, modified_parameters = list())
}
\arguments{
  \item{scenario_name}{The name of the scenario to
  retrieve}

  \item{modified_parameters}{A list of parameters to
  change. This will override the values from the stored
  scenario}
}
\description{
  This function is a poor way of storing pre-specified
  scenarios. This function is mainly a placeholder for a
  mechanism that will store specified scenarios in a DB
}

