\name{make_scenario_ui}
\alias{make_scenario_ui}
\title{This function produces the shiny components that can be rendered as a UI}
\usage{
  make_scenario_ui(scenario)
}
\arguments{
  \item{scenario}{The scenario to produce the ui from}
}
\description{
  This function produces the shiny components that can be
  rendered as a UI
}

