\name{plot_component}
\alias{plot_component}
\title{Plots a component of the system}
\usage{
  plot_component(plot_vars, component = "Strain")
}
\arguments{
  \item{plot_vars}{A solved system that has been formatted
  with \link{format_data}}

  \item{component}{A character vector that can take on any
  of the values (Strain, tCells or notMutation)}
}
\description{
  Plots a component of the system
}

