\name{rootFunc}
\alias{rootFunc}
\title{This function searched for roots in the system}
\usage{
  rootFunc(t, y, parms)
}
\arguments{
  \item{t}{time in the system}

  \item{y}{state variables}

  \item{parms}{The extra system parameters}
}
\description{
  There are three types of roots: 1) Treatment was changed
  2) A strain goes extinct 3) A mutation event happens
}

