\name{treatment_ui_from_scenario}
\alias{treatment_ui_from_scenario}
\title{Constructs the treatment part of the ui from a scenario}
\usage{
  treatment_ui_from_scenario(scenario)
}
\arguments{
  \item{scenario}{The scenario to produce the ui from}
}
\description{
  Constructs the treatment part of the ui from a scenario
}

