\name{compute_parameters}
\alias{compute_parameters}
\title{This function transforms the input parameters into a format that the equations can use}
\usage{
  compute_parameters(scenario_parameters)
}
\arguments{
  \item{scenario_parameters}{The parameters that specify
  the scenario}
}
\description{
  This function transforms the input parameters into a
  format that the equations can use
}

